아데오 에이전트 시스템/requirements.txt
gspread>=6.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
requests>=2.31.0
pyyaml>=6.0
openpyxl>=3.1.0
